requirements.txt
aiogram~=2.7
aiohttp>=3.6.2
aioredis==1.3.1
requests==2.26
python-dotenv~=0.18.0

tzlocal<3.0
Pillow~=8.0.1
environs~=9.3.3
APScheduler~=3.6.3